chore: Move tests to DatabaseImplTest.kt
